feat(ui): refactor of latency and minor fixes

update
